+++ hmri_create_MTProt.m
@@ -86,7 +86,7 @@ function [fR1, fR2s, fMT, fA, PPDw, PT1w, PMTw]  = hmri_create_MTProt(jobsubj) %
 
 flags = jobsubj.log.flags;
 flags.PopUp = false;
-hmri_log(sprintf('\t============ MPM PROCESSING - %s.m (%s) ============', mfilename, datestr(now)),flags);
+hmri_log(sprintf('\t============ MPM PROCESSING - %s.m (%s) ============', mfilename, datetime('now')),flags);
 
 % retrieves all required parameters for MPM processing
 mpm_params = get_mpm_params(jobsubj);
@@ -392,11 +392,18 @@ if mpm_params.proc.R2sOLS && any(mpm_params.estaticsR2s)
                 Ni = hmri_create_nifti(Pte0{ccon}, V_pdw(1), dt, ...
                     sprintf('%s fit to TE=0 for %sw images - %d echoes', mpm_params.R2s_fit_method, mpm_params.input(ccon).tag, length(mpm_params.input(ccon).TE)));
                                 
-                % set metadata
+                % store processing history in metadata
                 input_files = mpm_params.input(ccon).fnam;
                 Output_hdr = init_mpm_output_metadata(input_files, mpm_params);
                 Output_hdr.history.output.imtype = Ni.descrip;
                 Output_hdr.history.output.units = 'a.u.';
+
+                % copy acquisition metadata so extrapolated data could be fed back into 
+                % the toolbox if needed
+                Output_hdr.acqpar = struct('RepetitionTime',mpm_params.input(ccon).TR, ...
+                    'EchoTime',0, ...
+                    'FlipAngle',mpm_params.input(ccon).fa);
+
                 set_metadata(Pte0{ccon},Output_hdr,mpm_params.json);
                 
                 % re-load the updated NIFTI file (in case extended header has
@@ -952,7 +959,7 @@ spm_jsonwrite(fullfile(supplpath,'hMRI_map_creation_mpm_params.json'),mpm_params
 
 spm_progress_bar('Clear');
 
-hmri_log(sprintf('\t============ MPM PROCESSING: completed (%s) ============', datestr(now)),mpm_params.nopuflags);
+hmri_log(sprintf('\t============ MPM PROCESSING: completed (%s) ============', datetime('now')),mpm_params.nopuflags);
 
 
 end